performance: rescoped email methods to be in users.ajax, adjusted tests
