+++ books/urls.py
@@ -18,7 +18,4 @@ urlpatterns = [
 
     # ajax
     path("send-book-ajax/<uuid:pk>/", ajax.send_book_ajax, name="send_book_ajax"),
-    path("add-email/", ajax.add_email, name="add_email"),
-    path('delete-email/<int:pk>/', ajax.delete_email, name='delete_email'),
-    path("toggle-translate-email/<int:pk>/", ajax.toggle_translate_email, name="toggle_translate_email")
 ]
